Move CatonFixture in its own bazel module
