+++ canton/BUILD.bazel
@@ -1,0 +1,33 @@
+# Copyright (c) 2023 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
+# SPDX-License-Identifier: Apache-2.0
+
+load(
+    "//bazel_tools:scala.bzl",
+    "da_scala_library",
+    "lf_scalacopts_stricter",
+)
+
+java_binary(
+    name = "canton",
+    main_class = "com.digitalasset.canton.CantonCommunityApp",
+    visibility = ["//visibility:public"],
+    runtime_deps = [
+        # Do not sort
+        "@canton//:lib",
+    ],
+)
+
+# We replace engine and archive classes in canton
+# This should be used for testing only
+java_binary(
+    name = "canton-patched",
+    main_class = "com.digitalasset.canton.CantonCommunityApp",
+    visibility = ["//visibility:public"],
+    runtime_deps = [
+        # The following prevents buildifier from sorting the deps
+        # Do not sort
+        "//daml-lf/engine",
+        "//daml-lf/archive:daml_lf_archive_reader",
+        "@canton//:lib",
+    ],
+)
