+++ canton/it-lib/src/main/com/daml/CantonFixture.scala
@@ -1,7 +1,7 @@
 // Copyright (c) 2023 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package com.daml.lf
+package com.daml
 package integrationtest
 
 import com.daml.bazeltools.BazelRunfiles._
@@ -28,20 +28,6 @@ import java.util.UUID
 
 @scala.annotation.nowarn("msg=match may not be exhaustive")
 object CantonFixture {
-
-  final case class CompiledDar(
-      mainPkg: Ref.PackageId,
-      compiledPackages: PureCompiledPackages,
-  )
-
-  def readDar(
-      path: Path,
-      compilerConfig: speedy.Compiler.Config = speedy.Compiler.Config.Dev,
-  ): CompiledDar = {
-    val dar = archive.DarDecoder.assertReadArchiveFromFile(path.toFile)
-    val pkgs = PureCompiledPackages.assertBuild(dar.all.toMap, compilerConfig)
-    CompiledDar(dar.main._1, pkgs)
-  }
 
   private[integrationtest] lazy val List(serverCrt, serverPem, caCrt, clientCrt, clientPem) =
     List("server.crt", "server.pem", "ca.crt", "client.crt", "client.pem").map { src =>
@@ -99,7 +85,7 @@ trait CantonFixtureWithResource[A]
     implicit val resourceContext: ResourceContext = ResourceContext(system.dispatcher)
     new OwnedResource[ResourceContext, (Vector[Port], A)](
       for {
-        ports <- CantonRunner.run(config, cantonTmpDir, logger)
+        ports <- CantonRunner.run(config, cantonTmpDir, logger, darFiles)
         additional <- makeAdditionalResource(ports)
       } yield (ports, additional),
       acquisitionTimeout = 2.minute,
@@ -108,8 +94,6 @@ trait CantonFixtureWithResource[A]
   }
 
   lazy val config = CantonConfig(
-    applicationId = applicationId,
-    darFiles = darFiles,
     authSecret = authSecret,
     devMode = devMode,
     nParticipants = nParticipants,
@@ -137,19 +121,19 @@ trait CantonFixtureWithResource[A]
       token: Option[String] = None,
       maxInboundMessageSize: Int = 64 * 1024 * 1024,
   )(implicit ec: ExecutionContext): Future[LedgerClient] =
-    config.ledgerClient(ports.head, token, maxInboundMessageSize)
+    config.ledgerClient(ports.head, token, applicationId, maxInboundMessageSize)
 
   final protected def defaultLedgerClientWithoutId(
       token: Option[String] = None,
       maxInboundMessageSize: Int = 64 * 1024 * 1024,
   )(implicit ec: ExecutionContext): LedgerClientWithoutId =
-    config.ledgerClientWithoutId(ports.head, token, maxInboundMessageSize)
+    config.ledgerClientWithoutId(ports.head, token, applicationId, maxInboundMessageSize)
 
   final protected def ledgerClients(
       token: Option[String] = None,
       maxInboundMessageSize: Int = 64 * 1024 * 1024,
   )(implicit ec: ExecutionContext): Future[Vector[LedgerClient]] =
-    Future.traverse(ports)(config.ledgerClient(_, token, maxInboundMessageSize))
+    Future.traverse(ports)(config.ledgerClient(_, token, applicationId, maxInboundMessageSize))
 
 }
