+++ daml-script/export/integration-tests/reproduces-transactions/BUILD.bazel
@@ -33,8 +33,8 @@ da_scala_test(
     deps = [
         "//:sdk-version-scala-lib",
         "//bazel_tools/runfiles:scala_runfiles",
+        "//canton/it-lib",
         "//daml-lf/archive:daml_lf_archive_reader",
-        "//daml-lf/integration-test-lib",
         "//daml-lf/interpreter",
         "//daml-lf/language",
         "//daml-script/export",
