+++ language-support/java/codegen/BUILD.bazel
@@ -410,8 +410,8 @@ da_scala_test(
     deps = [
         ":ledger-tests-model.jar",
         "//bazel_tools/runfiles:scala_runfiles",
+        "//canton/it-lib",
         "//daml-lf/data",
-        "//daml-lf/integration-test-lib",
         "//language-support/java/bindings:bindings-java",
         "//language-support/scala/bindings-akka",
         "//ledger-api/rs-grpc-bridge",
