+++ language-support/java/codegen/src/ledger-tests/scala/com/digitalasset/TestUtil.scala
@@ -4,6 +4,7 @@
 package com.daml
 
 import com.daml.bazeltools.BazelRunfiles
+import com.daml.integrationtest.CantonFixture
 import com.daml.ledger.api.testing.utils.SuiteResourceManagementAroundAll
 import com.daml.ledger.api.v1.{ActiveContractsServiceGrpc, CommandServiceGrpc}
 import com.daml.ledger.api.v1.ActiveContractsServiceOuterClass.GetActiveContractsResponse
@@ -12,7 +13,6 @@ import com.daml.ledger.client.LedgerClient
 import com.daml.ledger.javaapi.data
 import com.daml.ledger.javaapi.data.{codegen => jcg, _}
 import com.daml.ledger.javaapi.data.codegen.HasCommands
-import com.daml.lf.integrationtest.CantonFixture
 import com.google.protobuf.Empty
 import io.grpc.Channel
 import org.scalatest.{Assertion, Suite}
