+++ language-support/scala/codegen-sample-app/BUILD.bazel
@@ -138,8 +138,8 @@ da_scala_test(
     deps = [
         ":daml-lf-codegen-sample-app",
         ":daml-lf-codegen-sample-app-testing",
+        "//canton/it-lib",
         "//daml-lf/data",
-        "//daml-lf/integration-test-lib",
         "//language-support/scala/bindings",
         "//language-support/scala/bindings-akka",
         "//language-support/scala/codegen-testing",
