+++ ledger/ledger-api-client/BUILD.bazel
@@ -99,8 +99,8 @@ da_scala_test_suite(
     deps = [
         ":ledger-api-client",
         "//bazel_tools/runfiles:scala_runfiles",
+        "//canton/it-lib",
         "//daml-lf/engine",
-        "//daml-lf/integration-test-lib",
         "//daml-lf/transaction",
         "//language-support/scala/bindings",
         "//ledger-api/rs-grpc-bridge",
