+++ ledger/ledger-api-client/src/it/scala/com/digitalasset/ledger/client/LedgerClientAuthIT.scala
@@ -4,6 +4,7 @@
 package com.daml.ledger.client
 
 import com.daml.grpc.GrpcException
+import com.daml.integrationtest.CantonFixture
 import com.daml.jwt.JwtSigner
 import com.daml.jwt.domain.DecodedJwt
 import com.daml.ledger.api.auth.AuthServiceJWTCodec
@@ -13,7 +14,6 @@ import com.daml.ledger.client.configuration.{
   LedgerIdRequirement,
   CommandClientConfiguration,
 }
-import com.daml.lf.integrationtest.CantonFixture
 import org.scalatest.Inside
 import org.scalatest.matchers.should.Matchers
 import org.scalatest.wordspec.AsyncWordSpec
