+++ navigator/backend/BUILD.bazel
@@ -182,9 +182,9 @@ da_scala_test_suite(
         ":navigator-tests-library",
         ":test-resources",
         "//bazel_tools/runfiles:scala_runfiles",
+        "//canton/it-lib",
         "//daml-lf/api-type-signature",
         "//daml-lf/data",
-        "//daml-lf/integration-test-lib",
         "//daml-lf/transaction",
         "//language-support/scala/bindings",
         "//ledger-api/rs-grpc-bridge",
