+++ triggers/service/BUILD.bazel
@@ -190,11 +190,11 @@ da_scala_library(
         ":trigger-service",
         ":trigger-service-binary-ce",
         "//bazel_tools/runfiles:scala_runfiles",
+        "//canton/it-lib",
         "//daml-lf/archive:daml_lf_1.dev_archive_proto_java",
         "//daml-lf/archive:daml_lf_archive_reader",
         "//daml-lf/data",
         "//daml-lf/engine",
-        "//daml-lf/integration-test-lib",
         "//daml-lf/interpreter",
         "//language-support/scala/bindings-akka",
         "//ledger-api/rs-grpc-bridge",
@@ -258,10 +258,10 @@ da_scala_test_suite(
         ":trigger-service",
         ":trigger-service-tests",
         "//bazel_tools/runfiles:scala_runfiles",
+        "//canton/it-lib",
         "//daml-lf/archive:daml_lf_1.dev_archive_proto_java",
         "//daml-lf/archive:daml_lf_archive_reader",
         "//daml-lf/data",
-        "//daml-lf/integration-test-lib",
         "//daml-lf/interpreter",
         "//language-support/scala/bindings-akka",
         "//ledger-api/rs-grpc-bridge",
