+++ triggers/tests/BUILD.bazel
@@ -21,6 +21,7 @@ da_scala_library(
         "src/test/scala/com/digitalasset/daml/lf/engine/trigger/test/AbstractFuncTests.scala",
         "src/test/scala/com/digitalasset/daml/lf/engine/trigger/test/AbstractTriggerTest.scala",
         "src/test/scala/com/digitalasset/daml/lf/engine/trigger/test/AbstractTriggerTestWithCanton.scala",
+        "src/test/scala/com/digitalasset/daml/lf/engine/trigger/test/CompiledDar.scala",
     ],
     scala_deps = [
         "@maven//:com_typesafe_akka_akka_stream",
@@ -33,9 +34,9 @@ da_scala_library(
     ],
     deps = [
         "//bazel_tools/runfiles:scala_runfiles",
+        "//canton/it-lib",
         "//daml-lf/archive:daml_lf_archive_reader",
         "//daml-lf/data",
-        "//daml-lf/integration-test-lib",
         "//daml-lf/interpreter",
         "//daml-lf/language",
         "//daml-lf/transaction",
@@ -149,10 +150,10 @@ EOF
             deps = [
                 ":test-utils",
                 "//bazel_tools/runfiles:scala_runfiles",
+                "//canton/it-lib",
                 "//daml-lf/archive:daml_lf_archive_reader",
                 "//daml-lf/data",
                 "//daml-lf/engine",
-                "//daml-lf/integration-test-lib",
                 "//daml-lf/interpreter",
                 "//daml-lf/language",
                 "//language-support/scala/bindings",
@@ -199,10 +200,10 @@ EOF
             deps = [
                 ":test-utils",
                 "//bazel_tools/runfiles:scala_runfiles",
+                "//canton/it-lib",
                 "//daml-lf/archive:daml_lf_archive_reader",
                 "//daml-lf/data",
                 "//daml-lf/engine",
-                "//daml-lf/integration-test-lib",
                 "//daml-lf/interpreter",
                 "//daml-lf/language",
                 "//language-support/scala/bindings",
@@ -262,8 +263,8 @@ da_scala_library(
     ],
     deps = [
         ":test-utils",
+        "//canton/it-lib",
         "//daml-lf/data",
-        "//daml-lf/integration-test-lib",
         "//daml-lf/interpreter",
         "//daml-lf/transaction",
         "//daml-script/converter",
@@ -304,10 +305,10 @@ da_scala_test_suite(
         ":test-utils",
         ":trigger-simulation-lib",
         "//bazel_tools/runfiles:scala_runfiles",
+        "//canton/it-lib",
         "//daml-lf/archive:daml_lf_archive_reader",
         "//daml-lf/data",
         "//daml-lf/engine",
-        "//daml-lf/integration-test-lib",
         "//daml-lf/interpreter",
         "//daml-lf/language",
         "//daml-lf/transaction",
@@ -358,10 +359,10 @@ da_scala_test_suite(
         ":test-utils",
         ":trigger-simulation-lib",
         "//bazel_tools/runfiles:scala_runfiles",
+        "//canton/it-lib",
         "//daml-lf/archive:daml_lf_archive_reader",
         "//daml-lf/data",
         "//daml-lf/engine",
-        "//daml-lf/integration-test-lib",
         "//daml-lf/interpreter",
         "//daml-lf/language",
         "//daml-lf/transaction",
