+++ triggers/tests/src/test/scala/com/digitalasset/daml/lf/engine/trigger/test/AbstractTriggerTestWithCanton.scala
@@ -9,6 +9,7 @@ package test
 import java.util.UUID
 import akka.stream.scaladsl.{Sink, Source}
 import com.daml.bazeltools.BazelRunfiles
+import com.daml.integrationtest.CantonFixture
 import com.daml.ledger.api.refinements.ApiTypes.Party
 import com.daml.ledger.api.v1.command_service.SubmitAndWaitRequest
 import com.daml.ledger.api.v1.commands.{Command, CreateCommand, ExerciseCommand, _}
@@ -24,7 +25,6 @@ import com.daml.ledger.client.configuration.{
 }
 import com.daml.lf.data.Ref._
 import com.daml.lf.engine.trigger.TriggerRunnerConfig.DefaultTriggerRunnerConfig
-import com.daml.lf.integrationtest.CantonFixture
 import com.daml.lf.speedy.SValue
 import com.daml.lf.speedy.SValue._
 import org.scalatest._
@@ -37,8 +37,6 @@ import scala.util.{Try, Success, Failure}
 // TODO: once test migration work has completed, rename this trait to AbstractTriggerTest
 trait AbstractTriggerTestWithCanton extends CantonFixture {
   self: Suite =>
-
-  import CantonFixture._
 
   private[this] lazy val darFile =
     Try(BazelRunfiles.requiredResource("triggers/tests/acs.dar").toPath) match {
@@ -76,7 +74,7 @@ trait AbstractTriggerTestWithCanton extends CantonFixture {
   protected def triggerRunnerConfiguration: TriggerRunnerConfig = DefaultTriggerRunnerConfig
 
   protected val CompiledDar(packageId, compiledPackages) =
-    readDar(darFile.merge, speedy.Compiler.Config.Dev)
+    CompiledDar.read(darFile.merge, speedy.Compiler.Config.Dev)
 
   protected def getRunner(
       client: LedgerClient,
@@ -234,4 +232,5 @@ object AbstractTriggerTestWithCanton {
       commandsInFlight: SValue,
       config: SValue,
   )
+
 }
